fix(e2e): move the sequencer under tests/ so the runner image can load it
The last run never executed a test: vitest failed at config load with
`Could not resolve './sequencer.js'`.

`Dockerfile.runner` copies `tests/` wholesale but only named files from
the package root (tsconfig.json, vitest.config.ts, package.json). A
sequencer beside the config therefore resolves on a workstation and is
simply absent in the image — which is why a local type-check passed and
CI broke. `diagnostics.ts` works because it already lives under `tests/`.

Put the sequencer there too and say why in the file, so the next helper
module does not repeat it. `include` only matches `*.spec.ts`, so a
non-spec module under `tests/` is never collected as a test.

Verified by loading the config against a tree containing only the files
the runner copies, rather than trusting a local type-check again.
